requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.6.3
attrs==25.4.0
cachetools==4.2.2
certifi==2026.1.4
charset-normalizer==3.4.4
class-registry==2.1.2
colorama==0.4.6
distro==1.9.0
filters==1.3.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
openai==0.28.0
OpenAIAuth==3.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
regex==2025.11.3
requests==2.32.5
revChatGPT==6.8.6
rich==14.2.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
tiktoken==0.12.0
tls-client==1.0.1
tornado==6.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==2.1
urllib3==1.26.16
wcwidth==0.2.14
wheel==0.45.1
yarl==1.22.0
python-dotenv
